+++ src/Modules/Matrix/Matrix.Presentation/Endpoints/Matrix/GetNextPos/GetNextPosRequest.cs
@@ -2,6 +2,6 @@ namespace Matrix.Presentation.Endpoints.Matrix.GetNextPos;
 
 public sealed class GetNextPosRequest
 {
-    public int M { get; init; }
+    public short M { get; init; }
     public string ProfileAddr { get; init; } = null!;
 }
