+++ src/Modules/Matrix/Matrix.Presentation/Endpoints/Places/GetCount/GetCountRequest.cs
@@ -2,6 +2,6 @@ namespace Matrix.Presentation.Endpoints.Places.GetCount;
 
 public sealed class GetCountRequest
 {
-    public int M { get; init; }
+    public short M { get; init; }
     public string ProfileAddr { get; init; } = null!;
 }
