+++ src/Modules/Matrix/Matrix.Presentation/Endpoints/Places/GetRootPlace/GetRootPlaceRequest.cs
@@ -2,6 +2,6 @@ namespace Matrix.Presentation.Endpoints.Places.GetRootPlace;
 
 public sealed class GetRootPlaceRequest
 {
-    public int M { get; init; }
+    public short M { get; init; }
     public string ProfileAddr { get; init; } = null!;
 }
