+++ src/Modules/Matrix/Matrix.Presentation/Endpoints/Places/SearchPaces/SearchPacesRequest.cs
@@ -2,7 +2,7 @@ namespace Matrix.Presentation.Endpoints.Places.SearchPaces;
 
 public sealed class SearchPacesRequest
 {
-    public int M { get; init; }
+    public short M { get; init; }
     public string ProfileAddr { get; init; } = null!;
     public int Page { get; init; } = 1;
     public int PageSize { get; init; } = 20;
